Add metadata tagging to core
