+++ packages/mongodb-rag-core/package.json
@@ -83,6 +83,7 @@
     "@supercharge/promise-pool": "^3.2.0",
     "acquit": "^1.3.0",
     "acquit-require": "^0.1.1",
+    "ai": "^4.3.9",
     "braintrust": "^0.0.193",
     "common-tags": "^1",
     "deep-equal": "^2.2.3",
