+++ packages/mongodb-rag-core/src/mongoDbMetadata/index.ts
@@ -1,3 +1,4 @@
+export * from "./classifyMetadata";
 export * from "./products";
 export * from "./programmingLanguages";
 export * from "./tags";
