+++ packages/mongodb-rag-core/src/mongoDbMetadata/products.ts
@@ -322,6 +322,7 @@ export const mongoDbProducts = [
 ] as const satisfies MongoDbProduct[];
 
 export type MongoDbProductName = (typeof mongoDbProducts)[number]["name"];
+export type MongoDbProductId = (typeof mongoDbProducts)[number]["id"];
 export type MongoDbProductNameEnum = [
   MongoDbProductName,
   ...MongoDbProductName[]
